create new store test file, add default state const
